LGT/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.5.0
tqdm>=4.65.0
numpy>=1.24.0
lpips>=0.1.4
transformers>=4.30.0
ftfy>=6.1.1
